Add draft mode flag for multi-zone (#68997)
When an application is running in multi-zone API endpoints can have
differing draft mode IDs on the same domain so we need to avoid clearing
the original when we see a mismatch as that can break the current
applications draft mode. This is behind a flag initially to validate but
it should be safe to make the default in the future as an invalid draft
mode ID just gets ignored.

x-ref: [slack
thread](https://vercel.slack.com/archives/C052BQ8F9EJ/p1723844164102539)